+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- update verify tests for GET/POST split (`test_http.py`)
+
 ## [0.2.0] - 2026-08-24
 
 ### Added